+++ conference-java/src/main/java/learn/springpath/conference/service/SpeakerServiceImpl.java
@@ -5,6 +5,7 @@ import learn.springpath.conference.repository.SpeakerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.PostConstruct;
 import java.util.List;
 
 @Service("speakerService")
@@ -25,6 +26,11 @@ public class SpeakerServiceImpl implements SpeakerService {
     public void setSpeakerRepository(SpeakerRepository speakerRepository) {
         this.speakerRepository = speakerRepository;
         System.out.println("SpeakerServiceImpl setter");
+    }
+
+    @PostConstruct
+    private void initialize() {
+        System.out.println("initialize() method called after the constructors ");
     }
 
     @Override